moved logic of getting projects to app-sidebar from nav-projects
