+++ src/components/universal/sidebar/app-sidebar.tsx
@@ -8,6 +8,9 @@ import {
   PersonStanding,
 } from 'lucide-react'
 
+import { type Project, type SidebarProject } from '@/types/Projects'
+import { getProjectsByUserId } from '@/api/projects'
+
 import { NavMain } from '@/components/universal/sidebar/nav-main'
 import { NavProjects } from '@/components/universal/sidebar/nav-projects'
 import { NavCommunities } from '@/components/universal/sidebar/nav-communities'
@@ -53,6 +56,27 @@ export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
   const [notificationsInfo, setNotificationsInfo] = React.useState<NotificatonsInfo>({ count: 0 })
   const [notificationsInfoLoading, setNotificationsInfoLoading] =
     React.useState<boolean>(true)
+
+  const [projects, setProjects] = React.useState<SidebarProject[]>([])
+  React.useEffect(() => {
+    const fetchProjects = async () => {
+      if (!user?.user) return
+
+      setProjectsLoading(true)
+      const response = await getProjectsByUserId(user.user)
+      if (response.data) {
+        const sidebarProjects: SidebarProject[] = response.data.map(project => ({
+          id: project.project_id,
+          title: project.title,
+          url: `/projects/${project.project_id}`
+        }))
+        setProjects(sidebarProjects)
+      }
+      setProjectsLoading(false)
+    }
+
+    void fetchProjects()
+  }, [user?.user])
 
   // Fetch communities/use cookies
   React.useEffect(() => {
@@ -178,8 +202,6 @@ export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
         ],
       },
     ],
-    projects: [
-    ],
   }
 
   // MacOS
@@ -219,7 +241,7 @@ export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
       <SidebarContent>
         <NavMain items={data.navMain} />
         <NavProjects
-          projects={data.projects}
+          projects={projects}
           loading={projectsLoading}
         />
         <NavCommunities
